Finished #19, with a bit of work on #17 and #18
Set up communication from activity --> service.
Buttons to start, stop, lap, reset work perfectly.
Some thought should maybe given to what the end goal
for the UI should be.

Service mostly seems to work, but testing is rough
on the emulator since it doesn't generate gps data.
I'll need to start using it on my phone to test it
fully.

I'm going to close #17 as well, as the recorder seems
to generate well-formed xml, but I might suggest further
research into if this is the best option.