Settings menu is now up, and can be interacted with.
Unfortunately, it has a few bugs. While the setting can
be changed back and forth, it does not seem to be being
saved or loaded properly, as units are always in metric
regardless of whether the setting points to statute or
metric. That will be the subject of the next commit.

Also did some slight UI work to make the fields bigger.