Fixed the notification manager.
Fixed it so that it didn't crash the
phone on android 2.3. Also actually tested out the app on
a real phone with gps data; once I get a case for this i'll
do more extensive testing on an actual ride. It seems like my
write issue is fixed and that I'll be able to test more extensively
on my real phone now; this should spur development.