Decided to restart the web dev side of things using Django
This is for a few reasons
     - I don't understand ruby, and I don't feel like using it. Django should be
       easier to pick up, and since I don't know webdev anyways, it will be good
       to start a new project
     - While there are definitely a lot of files and lines of code in the old
       website, there wasn't actually much done. It shouldn't take much work to
       to get back up to the old level of functionality.
     - The only really useful feature was viewing rides as an overlay on a map
       tile. However, this was implemented using the Google Maps API, and we
       intend on using OSM. That being said, it is likely that we can utilize
       the Google Maps version for guidance to the OSM version. On the other
       hand, this is likely no better than it would be if we were just to use
       the old ruby code.
     - The old file structure was rather poor and actually included a lot of
       unnecessary files.

In general, I think it will be a much more productive decision in the long run
to start from scratch. One thing that might be difficult is collaboration with
the database, since I haven't set it up on a static IP. If we're going to have
more than just me working on the website, that will need to be done.

Since I imagine I'm going to be the only user of this service for a while, I
might hold off for a while on setting up user authentication and similar
services. This way I can focus on the real services and benefits of this website.